KOGITO-1546 XStreamUtils kie-soup refactoring
