+++ kie-soup-xstream/pom.xml
@@ -22,18 +22,18 @@
   <parent>
     <groupId>org.kie.soup</groupId>
     <artifactId>kie-soup-parent</artifactId>
-    <version>7.38.0-SNAPSHOT</version>
+    <version>7.37.0-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
 
-  <artifactId>kie-soup-commons</artifactId>
+  <artifactId>kie-soup-xstream</artifactId>
   <packaging>bundle</packaging>
 
-  <name>KIE Soup Commons</name>
-  <description>Collection of reusable (not depending on any other KIE Soup module) components for KIE Soup.</description>
+  <name>KIE Soup XStream</name>
+  <description>Collection of XStream utilities (not depending on any other KIE Soup module) for KIE Soup.</description>
 
   <properties>
-    <java.module.name>org.kie.soup.commons</java.module.name>
+    <java.module.name>org.kie.soup.xstream</java.module.name>
   </properties>
 
   <dependencies>
@@ -42,8 +42,8 @@
          Don't add dependencies from this module to any other UberFire or Errai module! See UF-131 and UF-135 for details. -->
 
     <dependency>
-      <groupId>javax.enterprise</groupId>
-      <artifactId>cdi-api</artifactId>
+      <groupId>com.thoughtworks.xstream</groupId>
+      <artifactId>xstream</artifactId>
       <scope>provided</scope>
     </dependency>
     <dependency>
@@ -71,11 +71,12 @@
         <artifactId>maven-bundle-plugin</artifactId>
         <configuration>
           <instructions>
-            <Bundle-SymbolicName>org.kie.soup.commons</Bundle-SymbolicName>
+            <Bundle-SymbolicName>org.kie.soup.xstream</Bundle-SymbolicName>
             <Export-Package>
-              org.kie.soup.commons.*
+              org.kie.soup.xstream.*
             </Export-Package>
             <Import-Package>
+              com.thoughtworks.xstream.*;resolution:=optional,
               *
             </Import-Package>
             <_removeheaders>Private-Package</_removeheaders>
