+++ kie-soup-xstream/src/main/java/org/kie/soup/xstream/AnyAnnotationTypePermission.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kie.soup.commons.xstream;
+package org.kie.soup.xstream;
 
 import com.thoughtworks.xstream.annotations.XStreamAlias;
 import com.thoughtworks.xstream.annotations.XStreamAliasType;
