+++ kie-soup-xstream/src/main/java/org/kie/soup/xstream/LocalDateXStreamConverter.java
@@ -13,13 +13,10 @@
  * limitations under the License.
 */
 
-package org.kie.soup.commons.xstream;
+package org.kie.soup.xstream;
 
 import java.time.DateTimeException;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeFormatterBuilder;
-import java.time.temporal.ChronoField;
+import java.time.LocalDate;
 
 import com.thoughtworks.xstream.converters.Converter;
 import com.thoughtworks.xstream.converters.MarshallingContext;
@@ -32,37 +29,28 @@ import com.thoughtworks.xstream.io.HierarchicalStreamWriter;
  *
  * @see <a href="https://github.com/x-stream/xstream/issues/75">XStream#75</a>
  */
-public class LocalDateTimeXStreamConverter implements Converter {
-
-    private final DateTimeFormatter formatter;
-
-    public LocalDateTimeXStreamConverter() {
-        formatter = new DateTimeFormatterBuilder()
-                .appendPattern( "uuuu-MM-dd'T'HH:mm:ss" )
-                .appendFraction( ChronoField.NANO_OF_SECOND, 0, 9, true )
-                .toFormatter();
-    }
+public class LocalDateXStreamConverter implements Converter {
 
     @Override
-    public void marshal( Object localDateTimeObject, HierarchicalStreamWriter writer, MarshallingContext context ) {
-        LocalDateTime localDateTime = (LocalDateTime) localDateTimeObject;
-        writer.setValue( formatter.format( localDateTime ) );
+    public void marshal( Object localDateObject, HierarchicalStreamWriter writer, MarshallingContext context ) {
+        LocalDate localDate = (LocalDate) localDateObject;
+        writer.setValue( localDate.toString() );
     }
 
     @Override
     public Object unmarshal( HierarchicalStreamReader reader, UnmarshallingContext context ) {
-        String localDateTimeString = reader.getValue();
+        String localDateString = reader.getValue();
         try {
-            return LocalDateTime.from( formatter.parse( localDateTimeString ) );
+            return LocalDate.parse( localDateString );
         } catch ( DateTimeException e ) {
-            throw new IllegalStateException( "Failed to convert string (" + localDateTimeString + ") to type ("
-                    + LocalDateTime.class.getName() + ")." );
+            throw new IllegalStateException( "Failed to convert string (" + localDateString + ") to type ("
+                    + LocalDate.class.getName() + ")." );
         }
     }
 
     @Override
     public boolean canConvert( Class type ) {
-        return LocalDateTime.class.isAssignableFrom( type );
+        return LocalDate.class.isAssignableFrom( type );
     }
 
 }
