+++ kie-soup-xstream/src/main/java/org/kie/soup/xstream/XStreamUtils.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.kie.soup.commons.xstream;
+package org.kie.soup.xstream;
 
 import java.util.function.Function;
